Complete pull request
We only want to show active bonds in the list for confiscation.
The Inactive bond label is not necessary in the vote result as a
confiscated bond is inactive anyway. Intention was probably to show it
as inactive in the proposal display but they will be filtered out
anyway there.