Reduce tx broadcast timeout
Currently there is a bug in BitcoinJ causing the timeout at all BSQ
transactions.
It is because BitcoinJ does not handle confidence object correctly in
case as tx got altered after the
Wallet.complete() method is called which is the case for all BSQ txs.
We will work on a fix for that but that
will take more time. In the meantime we reduce the timeout to 5
seconds to avoid that the trade protocol runs
into a timeout when using BSQ for trade fee.