Fix issues with DAO full mode in preferences
There was missing the block notify port as well as the options have
not been applied correctly which led to the situation that when you
run the same application in full mode and later in lite mode that
data got persisted.

Added also to the footer a note if full mode is running.