Add duration log for DaoStateMonitoringService
Creating the daoStateHash is quite expensive (about 20-30 ms/block).
As this happens while parsing it delays the parsing time (parsing is
very fast). We persist t the hashes so it is only done for new blocks.
For 1 month of blocks (4000) it would take about 80-120 seconds.
It is not blocking the user thread as it is done per block and those
are parsed one after another.