Fix proposal table hide when switching to another screen and back
Do not remove the proposals table when VoteResultView is deactivated, as
the cycle selection which triggered its display is still active upon
switching to another screen and back again. This fixes #3269.

Also remove a spurious 'GUIUtil.removeChildrenFromGridPaneRows' call
when a proposal result is selected, as it is no longer needed since
switching from an inline votes table to a popup.