Revert "Apply kotlin plugin and convert one unused class to kotlin"
This reverts commit fc1c16ca3c0c19a1ff2c7a6c62daeba670aa4880 because
Kotlin compilation slows down the build, was applied too broadly to all
modules instead of just the one that needed it, and most importantly
because we never actually went ahead with converting anything of
importance to Kotlin. The commit being reverted was basically a demo,
converting a single test type to show what kind of difference it would
make.