Delete PersistableHashMap & simplify PersistableList
Remove unused functionality to pass a custom 'toProto' function via a
constructor or setter, as it was unused and all subclasses override the
toProtoMessage() method anyway. In this way, the method may be removed
and the class (+ UserThreadMappedPersistableList) made abstract.

Also it appears PersistableHashMap was never used, so remove it.