Fix an issue whereby the unread chat message count was not cleared properly
The trader chat view can be opened either by clicking on the chat icon
of a trade in the list, or clicking on "OPEN TRADER CHAT" button for the
currently displayed trade.  In the latter case, the count of new messages
displayed on the chat icon was not cleared even though trader chat
was shown.  The solution is to move the scope of the routine that
updates trader chat message count from within the table cell handler
to the PendingTradesView class so that it can be called whenever
onChat() is invoked (i.e. when the Trader Chat screen is shown).

Clicking on the trader chat icon of a trade that is not selected
should select that row in the trade list. It does not. This causes
confusion as it gives misleading cues as to which trader you are
communicating with.
Issue a call to select the row of the button when clicked.
Fixes #4172