Warning when restoring wallet seedwords without date
When the user restores a BIP39 wallet from seed words there is
an optional 'wallet date' field.  If the user does not enter that
field the chain scan will be a lengthy process.
Issue a popup message to the user in this case, informing them
that the synch will take a long time and can be reduced by either
specifying the wallet creation date or the date when Bisq introduced
BIP39 seed words.

Fixes #3766