Fix popup window ordering problem
There is a race condition in PopupManager when with a stack of 2 popups,
the top popup is being replaced by another top-level popup. The base
layer popup ends up on top of the z-order, obscuring the correct one.

The race condition in PopupManager was within the onHidden() method
which invokes the next window in the queue.  The call to displayNext()
needs to be asynchronous so that a newer popup can take precedence.

Fixes #4293