Remove hacked method tests
Although JUnit is geared towards running randomly ordered unit
tests, we can use it to write API tests to be run in a pre-defined
order.  This is necessary because API tests will be functional
(narrow to broad) and end-to-end;  we want to be able to write
@Test methods that run in @Order.

The next commit will add a JUnit OrderedRunner class, and an @Order
annotation.