Read numRequiredConfirmations from autoConfirmSettings just in time.
This allows the user to change required confirmation and it has impact
on open requests. Will be read at parsing the result.

Changed also the text field to not listen on text change but on focus
out so that the values are only updated once the user has left the
field and the value is valid. Otherwise small numbers like 1 might be
written during typing. The other values are only read at request start
and changes will have no impact on already started requests.

Also applies previous value in case the user added an invalid value at
focus out.

- Made display text shorter and added a tooltip.
- Allow very high numbers in devmode as max conf value