Upgrade apache-commons-lang3 to v11 to avoid jdeps issue
Updated the commons-lang3 version to the latest one, to avoid the existence of multiple versions used by different projects.

Before, v3.8 was used in most projects and v3.9 was brought as a transitive dependency in pricenode. So both were used by different projects. With this commit, all projects use v3.11, since it is explicitly defined in build.gradle.

This resolves an exception thrown by jdeps when trying to list the module dependencies of the desktop-vXYZ-SNAPSHOT-all.jar.