Allow user-defined block explorers
In Settings/General Preferences the Bitcoin Explorer and Bisq Explorer may
now be user defined in addition to the hard coded block explorer choices.

This implementation takes advantage of the existing Protobuf implementation
that stores a BlockExplorer Name/UrlTx/UrlAddress for BTC and for BSQ.

A new popup dialog is implemented which allows editing of the explorer
details (name, txUrl, and addressUrl).  Any of the system-defined
explorer details may be selected into the edit fields thus allowing
them to be used as templates, customized or an entirely new explorer to
be used.

If the user switches choice from a customized back to a pre-defined
explorer then the customized settings are gone; this is because the
Protobuf storage only allows saving one selection.

Fixes #4371
Closes #4634