Remove default fee rate from Bisq startup screen
The display of fee rate on the startup screen is misleading because
at startup a fee has not been received from the mempool fee service
and instead the default value of 50 sats is shown.

This change causes the fee rate not to be displayed until a response
has been received from the service.

Fixes #4765